requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
celery-progress==0.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
configparser==6.0.0
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-gravatar2==1.4.4
django-timezone-field==6.0.1
exceptiongroup==1.1.3
greenlet==3.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ImageHash==4.3.1
kombu==5.3.2
numpy==1.26.0
oauthlib==3.2.1
objectrest==2.0.0
opencv-python-headless==4.8.1.78
packaging==21.3
Pillow==10.0.1
PlexAPI==4.15.4
prompt-toolkit==3.0.39
proxlist==0.5.1
pydantic==1.10.13
pymediainfo==6.0.1
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.4.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
tautulli==3.5.2.2120
tmdbv3api==1.9.0
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.8
woodchips==0.2.4